merge

+++ canal-admin/canal-admin-server/src/main/java/com/alibaba/otter/canal/admin/service/impl/NodeServerServiceImpl.java
@@ -5,17 +5,21 @@ import io.ebean.Query;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import org.apache.commons.lang.StringUtils;
 import org.springframework.stereotype.Service;
 
+import com.alibaba.otter.canal.admin.common.Threads;
 import com.alibaba.otter.canal.admin.common.exception.ServiceException;
 import com.alibaba.otter.canal.admin.connector.AdminConnector;
 import com.alibaba.otter.canal.admin.connector.SimpleAdminConnectors;
+import com.alibaba.otter.canal.admin.model.CanalConfig;
+import com.alibaba.otter.canal.admin.model.CanalInstanceConfig;
 import com.alibaba.otter.canal.admin.model.NodeServer;
+import com.alibaba.otter.canal.admin.model.Pager;
 import com.alibaba.otter.canal.admin.service.NodeServerService;
 
 /**
@@ -31,7 +35,7 @@ public class NodeServerServiceImpl implements NodeServerService {
         int cnt = NodeServer.find.query()
             .where()
             .eq("ip", nodeServer.getIp())
-            .eq("admin_port", nodeServer.getAdminPort())
+            .eq("adminPort", nodeServer.getAdminPort())
             .findCount();
         if (cnt > 0) {
             throw new ServiceException("节点信息已存在");
@@ -48,25 +52,39 @@ public class NodeServerServiceImpl implements NodeServerService {
         int cnt = NodeServer.find.query()
             .where()
             .eq("ip", nodeServer.getIp())
-            .eq("admin_port", nodeServer.getAdminPort())
+            .eq("adminPort", nodeServer.getAdminPort())
             .ne("id", nodeServer.getId())
             .findCount();
         if (cnt > 0) {
             throw new ServiceException("节点信息已存在");
         }
 
-        nodeServer.update("name", "ip", "admin_port", "tcp_port", "metric_port");
+        nodeServer.update("name", "ip", "adminPort", "tcpPort", "metricPort", "clusterId");
     }
 
     public void delete(Long id) {
         NodeServer nodeServer = NodeServer.find.byId(id);
         if (nodeServer != null) {
+            // 判断是否存在实例
+            int cnt = CanalInstanceConfig.find.query().where().eq("serverId", nodeServer.getId()).findCount();
+            if (cnt > 0) {
+                throw new ServiceException("当前Server下存在Instance配置, 无法删除");
+            }
+
+            // 同时删除配置
+            CanalConfig canalConfig = CanalConfig.find.query().where().eq("serverId", id).findOne();
+            if (canalConfig != null) {
+                canalConfig.delete();
+            }
+
             nodeServer.delete();
         }
     }
 
-    public List<NodeServer> findList(NodeServer nodeServer) {
+    private Query<NodeServer> getBaseQuery(NodeServer nodeServer) {
         Query<NodeServer> query = NodeServer.find.query();
+        query.fetch("canalCluster", "name").setDisableLazyLoading(true);
+
         if (nodeServer != null) {
             if (StringUtils.isNotEmpty(nodeServer.getName())) {
                 query.where().like("name", "%" + nodeServer.getName() + "%");
@@ -74,18 +92,47 @@ public class NodeServerServiceImpl implements NodeServerService {
             if (StringUtils.isNotEmpty(nodeServer.getIp())) {
                 query.where().eq("ip", nodeServer.getIp());
             }
-        }
-        query.order().asc("id");
-        List<NodeServer> nodeServers = query.findList();
-        if (nodeServers.isEmpty()) {
-            return nodeServers;
+            if (nodeServer.getClusterId() != null) {
+                if (nodeServer.getClusterId() == -1) {
+                    query.where().isNull("clusterId");
+                } else {
+                    query.where().eq("clusterId", nodeServer.getClusterId());
+                }
+            }
         }
 
-        ExecutorService executorService = Executors.newFixedThreadPool(nodeServers.size());
+        return query;
+    }
+
+    public List<NodeServer> findAll(NodeServer nodeServer) {
+        Query<NodeServer> query = getBaseQuery(nodeServer);
+        query.order().asc("id");
+        return query.findList();
+    }
+
+    public Pager<NodeServer> findList(NodeServer nodeServer, Pager<NodeServer> pager) {
+
+        Query<NodeServer> query = getBaseQuery(nodeServer);
+        Query<NodeServer> queryCnt = query.copy();
+
+        int count = queryCnt.findCount();
+        pager.setCount((long) count);
+
+        List<NodeServer> nodeServers = query.order()
+            .asc("id")
+            .setFirstRow(pager.getOffset().intValue())
+            .setMaxRows(pager.getSize())
+            .findList();
+        pager.setItems(nodeServers);
+
+        if (nodeServers.isEmpty()) {
+            return pager;
+        }
+
         List<Future<Boolean>> futures = new ArrayList<>(nodeServers.size());
         // get all nodes status
         for (NodeServer ns : nodeServers) {
-            futures.add(executorService.submit(() -> {
+            futures.add(Threads.executorService.submit(() -> {
                 boolean status = SimpleAdminConnectors.execute(ns.getIp(), ns.getAdminPort(), AdminConnector::check);
                 ns.setStatus(status ? "1" : "0");
                 return !status;
@@ -93,15 +140,13 @@ public class NodeServerServiceImpl implements NodeServerService {
         }
         futures.forEach(f -> {
             try {
-                f.get();
-            } catch (InterruptedException | ExecutionException e) {
+                f.get(3, TimeUnit.SECONDS);
+            } catch (TimeoutException | InterruptedException | ExecutionException e) {
                 // ignore
             }
         });
 
-        executorService.shutdownNow();
-
-        return nodeServers;
+        return pager;
     }
 
     public int remoteNodeStatus(String ip, Integer port) {
